Move Context down to separate package
This will allow using Context in client-endpoint-api and auth-api
