+++ client-core/build.gradle.kts
@@ -8,6 +8,7 @@ extra["displayName"] = "Smithy :: Java :: Client Core"
 extra["moduleName"] = "software.amazon.smithy.java.client-core"
 
 dependencies {
+    api(project(":context"))
     api(project(":core"))
     api(project(":auth-api"))
     api(project(":client-endpoint-api"))
