+++ client-http/src/main/java/software/amazon/smithy/java/runtime/client/http/JavaHttpClientTransport.java
@@ -13,8 +13,8 @@ import java.time.Duration;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Flow;
+import software.amazon.smithy.java.context.Context;
 import software.amazon.smithy.java.runtime.client.core.ClientTransport;
-import software.amazon.smithy.java.runtime.core.Context;
 import software.amazon.smithy.java.runtime.http.api.SmithyHttpRequest;
 import software.amazon.smithy.java.runtime.http.api.SmithyHttpResponse;
 import software.amazon.smithy.java.runtime.http.api.SmithyHttpVersion;
