+++ context/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    id("smithy-java.module-conventions")
+}
+
+description = "This module provides a typed identity based collection"
+
+extra["displayName"] = "Smithy :: Java :: Context"
+extra["moduleName"] = "software.amazon.smithy.java.context"
+
+dependencies {
+}
